feat: make stack.py leaner
