+++ manifests/example-dev/core-modules.yaml
@@ -109,12 +109,16 @@ parameters:
       deploy_external_dns: True # We deploy it unless set to False
       deploy_aws_ebs_csi: True # We deploy it unless set to False
       deploy_aws_efs_csi: True # We deploy it unless set to False
+      deploy_aws_fsx_csi: False # We deploy it unless set to False
       deploy_cluster_autoscaler: True # We deploy it unless set to False
       deploy_metrics_server: True # We deploy it unless set to False
       deploy_secretsmanager_csi: True # We deploy it unless set to False
+      deploy_external_secrets: False
       deploy_cloudwatch_container_insights_metrics: True # We deploy it unless set to False
       deploy_cloudwatch_container_insights_logs: True
       cloudwatch_container_insights_logs_retention_days: 7
+      deploy_amp: False
+      deploy_grafana_for_amp: False
       deploy_kured: False
       deploy_calico: False
       deploy_nginx_controller:
