+++ manifests/mlops/core-modules.yaml
@@ -56,9 +56,11 @@ parameters:
       deploy_external_dns: True # We deploy it unless set to False
       deploy_aws_ebs_csi: True # We deploy it unless set to False
       deploy_aws_efs_csi: True # We deploy it unless set to False
+      deploy_aws_fsx_csi: False # We deploy it unless set to False
       deploy_cluster_autoscaler: True # We deploy it unless set to False
       deploy_metrics_server: True # We deploy it unless set to False
       deploy_secretsmanager_csi: True # We deploy it unless set to False
+      deploy_external_secrets: False
       deploy_cloudwatch_container_insights_metrics: True # We deploy it unless set to False
       deploy_cloudwatch_container_insights_logs: True
       cloudwatch_container_insights_logs_retention_days: 7
