+++ modules/core/eks/network-policies/default-allow-kube-system.json
@@ -1,0 +1,17 @@
+{
+    "apiVersion": "networking.k8s.io/v1",
+    "kind": "NetworkPolicy",
+    "metadata": {
+        "name": "default-allow-kube-system",
+        "namespace": "kube-system"
+    },
+    "spec": {
+        "ingress": [
+            {}
+        ],
+        "podSelector": {},
+        "policyTypes": [
+            "Ingress"
+        ]
+    }
+}
