+++ modules/core/eks/network-policies/default-allow-kyverno.json
@@ -1,0 +1,17 @@
+{
+    "apiVersion": "networking.k8s.io/v1",
+    "kind": "NetworkPolicy",
+    "metadata": {
+        "name": "default-allow-kyverno",
+        "namespace": "kyverno"
+    },
+    "spec": {
+        "ingress": [
+            {}
+        ],
+        "podSelector": {},
+        "policyTypes": [
+            "Ingress"
+        ]
+    }
+}
