+++ modules/core/eks/network-policies/default-allow-tigera-operator.json
@@ -1,0 +1,17 @@
+{
+    "apiVersion": "networking.k8s.io/v1",
+    "kind": "NetworkPolicy",
+    "metadata": {
+        "name": "default-allow-tigera-operator",
+        "namespace": "tigera-operator"
+    },
+    "spec": {
+        "ingress": [
+            {}
+        ],
+        "podSelector": {},
+        "policyTypes": [
+            "Ingress"
+        ]
+    }
+}
