+++ modules/core/eks/network-policies/default-deny.json
@@ -1,0 +1,13 @@
+{
+  "apiVersion": "crd.projectcalico.org/v1",
+  "kind": "GlobalNetworkPolicy",
+  "metadata": {
+    "name": "default-deny-app-policy"
+  },
+  "spec": {
+    "selector": "all()",
+    "types": [
+      "Ingress"
+    ]
+  }
+}
